Implemented promote command in blueprints

+++ commands/utility.py
@@ -57,7 +57,7 @@ def archive(name: str, filetype: str):
             print(f"{name} does not exist in forge.")
 
 def promote(name: str):
-    if not os.path.exists(f"ideas/forge/{name}"):
+    if not os.path.exists(f"ideas/forge/{name}") and os.path.exists(f"ideas/blueprints/{name}.md"):
         os.makedirs(f"ideas/forge/{name}")
         shutil.move(f"ideas/blueprints/{name}.md", f"ideas/forge/README.md")
         print(f"{name} promoted from blueprints to forge.")
